passing commands from appropriate component
