+++ client/content/sites/allSites/allSitesComponent.tsx
@@ -1,10 +1,14 @@
 import { DetailsList, DetailsListLayoutMode } from "office-ui-fabric-react/lib/DetailsList";
 import * as React from "react";
+import { Link } from "react-router-dom";
 
+import ButtonType from "../common/header/commandBarButtons/enums/buttonType";
 import GeneralError from "../../../common/generalError/generalError";
-import HeaderFor from "../common/header/duck/operations/enums/headerFor";
 import IAllSitesProps from "./interfaces/IAllSitesProps";
 import IHeaderPayload from "../common/header/duck/operations/interfaces/IHeaderPayload";
+import IOpenRightPanelPayload from "../../../chrome/duck/actions/interfaces/IOpenRightPanelPayload";
+import ItemLocation from "../common/header/commandBarButtons/enums/itemLocation";
+import IToggleSwitchRightPanePayload from "../../common/rightPane/duck/actions/interfaces/IToggleSwitchRightPanePayload";
 import LoadingSpinner from "../../../common/loadingSpinner/loadingSpinner";
 import QueryStringUtils from "../../../utils/queryStringUtils";
 import siteDetailsListColumns from "../../common/detailsList/siteDetailsList/SiteDetailsListColumns";
@@ -13,8 +17,20 @@ export default class AllSitesComponent extends React.PureComponent<IAllSitesProp
     private headerPayload: IHeaderPayload = {
         locationPath: null,
         entityTitle: "Sites",
-        headerFor: HeaderFor.AllSites,
-        isUpdateCommands: true
+        commands: [
+            {
+                id: ButtonType.Add,
+                name: "Site",
+                itemLocation: ItemLocation.Left,
+                actionPayload: rightPanelData
+            },
+            {
+                id: ButtonType.Info,
+                name: null,
+                itemLocation: ItemLocation.Far,
+                actionPayload: rightPaneData
+            }
+        ]
     };
 
     public render(): JSX.Element {
@@ -63,3 +79,24 @@ export default class AllSitesComponent extends React.PureComponent<IAllSitesProp
         this.props.setHeader(this.headerPayload);
     }
 }
+
+const Test2: React.SFC<object> = (): JSX.Element => (
+    <div>
+        Click on the next button to navigate to site details
+        <Link to="/sites/909876"> NEXT</Link>
+    </div>
+);
+
+const rightPaneData: IToggleSwitchRightPanePayload = {
+    rightPaneId: "allSitesRightPaneId",
+    rightPaneHeaderText: "Right pane for all sites",
+    rightPaneContent: <div>This is the body of the right pane...</div>,
+    rightPaneFooterRender: (): JSX.Element => (<div>Footer for the right pane</div>)
+};
+
+const rightPanelData: IOpenRightPanelPayload = {
+    rightPanelId: "allSitesRightPanelId",
+    rightPanelHeaderText: "Test Panel",
+    rightPanelContent: <Test2 />,
+    rightPanelFooterRender: (): JSX.Element => (<div>footer of the panel</div>)
+};
