--- client/content/sites/common/header/duck/operations/enums/headerFor.ts
@@ -1,6 +1,0 @@
-enum HeaderFor {
-    AllSites,
-    SiteDetails
-}
-
-export default HeaderFor;
