+++ client/content/sites/common/header/duck/operations/headerSetHeader.ts
@@ -1,13 +1,11 @@
 import { Dispatch } from "redux";
 
-import getCommandsPayload from "./getCommandsPayload";
 import headerSetBreadcrumb from "../actions/headerSetBreadcrumb";
 import headerSetCommands from "../actions/headerSetCommands";
 import headerSetEntityTitle from "../actions/headerSetEntityTitle";
 import IHeaderPayload from "./interfaces/IHeaderPayload";
 import IHeaderSetBreadcrumbAction from "../actions/interfaces/IHeaderSetBreadcrumbAction";
 import IHeaderSetCommandsAction from "../actions/interfaces/IHeaderSetCommandsAction";
-import IHeaderSetCommandButtonsPayload from "../actions/interfaces/IHeaderSetCommandsPayload";
 import IHeaderSetEntityTitleAction from "../actions/interfaces/IHeaderSetEntityTitleAction";
 
 type Actions = IHeaderSetEntityTitleAction | IHeaderSetCommandsAction | IHeaderSetBreadcrumbAction;
@@ -28,12 +26,10 @@ export default function headerSetHeader(
             }));
         }
 
-        if (headerPayload.isUpdateCommands) {
-            const commandsPayload: IHeaderSetCommandButtonsPayload = getCommandsPayload(headerPayload.headerFor);
-
-            if (commandsPayload) {
-                dispatch(headerSetCommands(commandsPayload));
-            }
+        if (headerPayload.commands) {
+            dispatch(headerSetCommands({
+                buttonList: headerPayload.commands
+            }));
         }
     };
 }
