+++ client/content/sites/siteDetails/siteDetailsContainer.ts
@@ -6,6 +6,9 @@ import { createSelector, OutputSelector } from "reselect";
 import headerSetHeader from "../common/header/duck/operations/headerSetHeader";
 import IAppState from "../../../duck/interfaces/IAppState";
 import IHeaderPayload from "../common/header/duck/operations/interfaces/IHeaderPayload";
+import IJobData from "../data/duck/interfaces/IJobData";
+import IJobTileData from "../../common/jobSummaryList/interfaces/IJobTileData";
+import ISiteData from "../data/duck/interfaces/ISiteData";
 import ISiteDetailsProps from "./interfaces/ISiteDetailsProps";
 import ISiteDetailsPropsFromDispatch from "./interfaces/ISiteDetailsPropsFromDispatch";
 import ISiteDetailsPropsFromState from "./interfaces/ISiteDetailsPropsFromState";
@@ -15,9 +18,6 @@ import SiteDetailsComponent from "./siteDetailsComponent";
 import siteDetailsGetData from "./duck/operations/siteDetailsGetData";
 import SiteDetailsShim from "./shim/siteDetailsShim";
 import sitesToggleRightPane from "../duck/actions/sitesToggleRightPane";
-import IJobData from "../data/duck/interfaces/IJobData";
-import IJobTileData from "../../common/jobSummaryList/interfaces/IJobTileData";
-import ISiteData from "../data/duck/interfaces/ISiteData";
 
 type GetJobIdsFromSiteDetailsState = (state: IAppState) => string[];
 type GetJobDataFromSitesDataState = (state: IAppState) => IJobData;
